# lib/api/openapi.yaml
openapi: 3.0.3
info:
  title: RestaurantOS API
  description: |
    Full REST API for RestaurantOS v2 Restaurant Management System.
    All protected endpoints require a Bearer JWT token.
  version: 2.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:    { type: string }
                password: { type: string }
      responses:
        '200': { description: Login successful }
        '401': { description: Invalid credentials }
...
  /auth/logout:
    post:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Logout
      responses:
        '200': { description: Logged out }
  /auth/me:
    get:
      tags: [Auth]
      security: [{ BearerAuth: [] }]
      summary: Get current user
      responses:
        '200': { description: Current user data }
...
  /menu:
    get:
      tags: [Menu]
      security: [{ BearerAuth: [] }]
      summary: List menu items
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200': { description: Menu items list }
    post:
      tags: [Menu]
      security: [{ BearerAuth: [] }]
      summary: Create menu item (Admin/Manager)
      responses:
        '201': { description: Item created }
  /menu/categories:
    get:
      tags: [Menu]
      security: [{ BearerAuth: [] }]
      summary: List categories
      responses:
        '200': { description: Categories }
  /menu/{id}:
    put:
      tags: [Menu]
      security: [{ BearerAuth: [] }]
      summary: Update menu item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Item updated }
    delete:
      tags: [Menu]
      security: [{ BearerAuth: [] }]
      summary: Deactivate item (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Item deactivated }
  /orders:
    get:
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: Orders list }
    post:
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      summary: Create order (auto-deducts inventory)
      responses:
        '201': { description: Order created }
  /orders/{id}:
    get:
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      summary: Get single order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Order detail }
    delete:
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      summary: Cancel order (Admin/Manager)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Order cancelled }
  /orders/{id}/status:
    put:
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Status updated }
...
  /inventory:
    get:
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      summary: List inventory
      responses:
        '200': { description: Inventory list }
  /inventory/{id}:
    put:
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      summary: Update stock (Admin/Manager)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Inventory updated }
  /inventory/{id}/transactions:
    get:
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      summary: Get transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Transaction log }
  /customers:
    get:
      tags: [Customers]
      security: [{ BearerAuth: [] }]
      summary: List customers
      responses:
        '200': { description: Customers list }
    post:
      tags: [Customers]
      security: [{ BearerAuth: [] }]
      summary: Create customer
      responses:
        '201': { description: Customer created }
  /customers/{id}/orders:
    get:
      tags: [Customers]
      security: [{ BearerAuth: [] }]
      summary: Order history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Order history }
  /customers/{id}/loyalty:
    put:
      tags: [Customers]
      security: [{ BearerAuth: [] }]
      summary: Update loyalty points
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Points updated }
...